Add some static casts

+++ src/worldgen/Overworld.cpp
@@ -102,7 +102,7 @@ namespace Game3::WorldGen {
 				const Index col_max = col_min + CHUNK_SIZE;
 
 				pool.add([&, game_ptr, row_min, row_max, col_min, col_max](ThreadPool &, size_t) {
-					threadContext = {game_ptr, noise_seed - 1'000'000ul * row_min + col_min, row_min, row_max, col_min, col_max};
+					threadContext = {game_ptr, static_cast<uint_fast32_t>(noise_seed - 1'000'000ul * row_min + col_min), row_min, row_max, col_min, col_max};
 
 					std::vector<double> saved_noise((row_max - row_min) * (col_max - col_min));
 
@@ -242,7 +242,7 @@ namespace Game3::WorldGen {
 				// Compare with <, not <=
 				const Index col_max = col_min + CHUNK_SIZE;
 				pool.add([realm, &waiter, &get_biome, &perlin, &params, noise_seed, row_min, row_max, col_min, col_max](ThreadPool &, size_t) {
-					threadContext = {realm->getGame().shared_from_this(), noise_seed - 1'000'000ul * row_min + col_min, row_min, row_max, col_min, col_max};
+					threadContext = {realm->getGame().shared_from_this(), static_cast<uint_fast32_t>(noise_seed - 1'000'000ul * row_min + col_min), row_min, row_max, col_min, col_max};
 					for (Index row = row_min; row < row_max; ++row)
 						for (Index column = col_min; column < col_max; ++column)
 							get_biome(row, column).postgen(row, column, threadContext.rng, perlin, params);
